Most generic parts of orchestration split out into cadorchestrator
